requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
openai==1.12.0
packaging==23.2
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pygame==2.5.2
Pygments==2.17.2
requests==2.31.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
watchdog==4.0.0

